+++ cob_obstacle_distance/src/marker_shapes_interface.cpp
@@ -14,7 +14,7 @@
  * \note
  *   ROS stack name: cob_control
  * \note
- *   ROS package name: cob_collision_object_publisher
+ *   ROS package name: cob_obstacle_distance
  *
  * \author
  *   Author: Marco Bezzon, email: Marco.Bezzon@ipa.fraunhofer.de
@@ -26,7 +26,7 @@
  *   Important to note: increases a static class counter to ensure
  *   a unique ID for the rviz Marker publishing.
  ****************************************************************/
-#include "cob_collision_object_publisher/marker_shapes_interface.hpp"
+#include "cob_obstacle_distance/marker_shapes_interface.hpp"
 
 /* BEGIN IMarkerShape *******************************************************************************************/
 /// Interface class marking methods that have to be implemented in derived classes.
