+++ cob_obstacle_distance/srv/ObjectOfInterest.srv
