+++ cob_twist_controller/package.xml
@@ -22,7 +22,8 @@
   <build_depend>tf</build_depend>
   <build_depend>tf_conversions</build_depend>
   <build_depend>dynamic_reconfigure</build_depend>
-  <build_depend>cob_collision_object_publisher</build_depend>
+  <build_depend>cob_obstacle_distance</build_depend>
+  <build_depend>message_generation</build_depend>
   
   <run_depend>roscpp</run_depend>
   <run_depend>rospy</run_depend>
@@ -35,7 +36,8 @@
   <run_depend>tf</run_depend>
   <run_depend>tf_conversions</run_depend>
   <run_depend>dynamic_reconfigure</run_depend>
-  <run_depend>cob_collision_object_publisher</run_depend>
+  <run_depend>cob_obstacle_distance</run_depend>
+  <run_depend>message_runtime</run_depend>
 
   <export>
   </export>
